app/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
pymongo
python-dotenv
jinja2
langchain
langchain-google-genai
google-generativeai
pinecone
requests
pandas
numpy
matplotlib
seaborn
reportlab
textblob
aiohttp
asyncio-throttle
python-dateutil
pydantic-settings
websockets
yfinance
indstocks
nsetools
nsepython
pandas-datareader
mplfinance
ta
investpy

# Optional (for advanced visuals and research)
plotly
scikit-learn
nsepython
pandas-datareader
mplfinance
ta
investpy
email-validator
langchain-community